Separate Topology from Engine
