+++ src/examples/SubeventExample/SubeventExample.cc
@@ -4,10 +4,11 @@
 
 #include <JANA/JApplication.h>
 #include <JANA/JObject.h>
-#include <JANA/Engine/JSubeventArrow.h>
 #include <JANA/JEventSource.h>
 #include <JANA/JEventProcessor.h>
-#include "JANA/Engine/JTopologyBuilder.h"
+
+#include <JANA/Topology/JSubeventArrow.h>
+#include "JANA/Topology/JTopologyBuilder.h"
 
 
 struct MyInput : public JObject {
