+++ src/libraries/JANA/Topology/JEventMapArrow.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include <JANA/Engine/JPipelineArrow.h>
+#include <JANA/Topology/JPipelineArrow.h>
 
 class JEventPool;
 class JEventSource;
