+++ src/libraries/JANA/Topology/JEventSourceArrow.cc
@@ -5,7 +5,7 @@
 
 #include <JANA/JApplication.h>
 #include <JANA/JEventSource.h>
-#include <JANA/Engine/JEventSourceArrow.h>
+#include <JANA/Topology/JEventSourceArrow.h>
 #include <JANA/Utils/JEventPool.h>
