+++ src/libraries/JANA/Topology/JFoldArrow.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include <JANA/Engine/JArrow.h>
+#include <JANA/Topology/JArrow.h>
 #include <JANA/Utils/JEventPool.h>
 #include <JANA/Utils/JEventLevel.h>
