+++ src/libraries/JANA/Topology/JPipelineArrow.h
@@ -4,9 +4,9 @@
 
 #pragma once
 
-#include <JANA/Engine/JArrow.h>
-#include <JANA/Engine/JMailbox.h>
-#include <JANA/Engine/JPool.h>
+#include <JANA/Topology/JArrow.h>
+#include <JANA/Topology/JMailbox.h>
+#include <JANA/Topology/JPool.h>
 
 template <typename DerivedT, typename MessageT>
 class JPipelineArrow : public JArrow {
