+++ src/programs/unit_tests/QueueTests.cc
@@ -2,7 +2,7 @@
 // Copyright 2020, Jefferson Science Associates, LLC.
 // Subject to the terms in the LICENSE file found in the top-level directory.
 
-#include <JANA/Engine/JMailbox.h>
+#include <JANA/Topology/JMailbox.h>
 
 #include "catch.hpp"
